+++ Demo/Demo/PayPalVault/PayPalVaultViewModel/PayPalVaultViewModel.swift
@@ -26,7 +26,7 @@ class PayPalVaultViewModel: VaultViewModel {
             ssid: userSSID
         )
         client.createPayPalSession(
-            flowType: .billingWithoutPurchase,
+            sessionType: .vaultWithoutPurchase,
             userIdentity: resolvedUserIdentity,
             urlConfig: ShopperSessionURLConfigFactory.urlConfig,
             userAction: selectedUserAction
@@ -49,7 +49,7 @@ class PayPalVaultViewModel: VaultViewModel {
                     continuation.resume()
                 case .failure(let error):
                     if error == PayPalError.vaultCanceledError {
-                        self.state.paypalVaultTokenResponse = .idle
+                        self.state.paypalVaultTokenResponse = .error(message: "PayPal vault was canceled.")
                         continuation.resume()
                     } else {
                         self.state.paypalVaultTokenResponse = .error(message: error.localizedDescription)